+++ backend/src/models/diaryModel.js
@@ -2,14 +2,14 @@
 const db = require("../config/db");
 
 // 查询日记
-function listDiaries() {
+function listDiaries(userId) {
   // 将db.all这个callback风格异步转换成promise风格异步(使得可以使用async/await)
   return new Promise((resolve, reject) => {
     db.all(
-      `SELECT id, mood, title, content, image_url as imageUrl, created_at as createdAt,image_ratio as imageRatio
-      FROM diaries
-      ORDER BY id ASC`,
-      [],
+      `SELECT id,mood, title, content, image_url as imageUrl, created_at as createdAt,image_ratio as imageRatio
+      FROM diaries WHERE user_id = ?
+      ORDER BY created_at ASC`,
+      [userId],
       (err, rows) => {
         if (err) return reject(err);
         resolve(rows);
@@ -20,12 +20,12 @@ function listDiaries() {
 
 // 创建日记
 function createDiary(cardInfo) {
-  let { mood, title, content, imageUrl, imageRatio, createdAt } = cardInfo;
+  let { userId,mood, title, content, imageUrl, imageRatio, createdAt } = cardInfo;
   return new Promise((resolve, reject) => {
     db.run(
-      `INSERT INTO diaries (mood, title, content, image_url,image_ratio, created_at)
-      VALUES(?, ?, ?, ?,?, ?)`,
-      [mood, title, content, imageUrl, imageRatio, createdAt],
+      `INSERT INTO diaries (mood, title, content, image_url,image_ratio, created_at,user_id)
+      VALUES(?, ?, ?, ?,?, ?,?)`,
+      [mood, title, content, imageUrl, imageRatio, createdAt,userId],
       function (err) {
         if (err) return reject(err);
 
@@ -44,11 +44,11 @@ function createDiary(cardInfo) {
 }
 
 // 查找日记图片
-function findDiaryImageById(id) {
+function findDiaryImageById(id,userId) {
   return new Promise((resolve, reject) => {
     db.get(
-      `SELECT image_url as imageUrl FROM diaries WHERE id = ?`,
-      [id],
+      `SELECT image_url as imageUrl FROM diaries WHERE id = ? AND user_id = ?`,
+      [id,userId],
       (err, row) => {
         if (err) return reject(err);
 
@@ -59,14 +59,18 @@ function findDiaryImageById(id) {
 }
 
 // 移除日记
-function removeDiary(id) {
+function removeDiary(id,userId) {
   return new Promise((resolve, reject) => {
     db.run(
-      `DELETE FROM diaries WHERE id = ?`,
-      [id],
-      (err) => {
-        if (err) return reject(err);
-        resolve();
+      `DELETE FROM diaries WHERE id = ? AND user_id = ?`,
+      [id,userId],
+      function(err){
+        if(err){
+          return reject(err);
+        }
+
+        // 此处的this是SQL执行的结果对象，changes是影响的数据行数，成功删除一行数据就是1
+        resolve(this.changes);
       }
     )
   })
